Remove the bundled (out-of-date) copy of include.js
The Persona team currently recommends against bundling that file
because it's still subject to change and using an old copy could
break logins. The best thing for now is to load the latest one
from login.persona.org directly.

Making it possible for sites to self-host include.js is however
on the roadmap:

  https://github.com/mozilla/browserid-roadmap/issues/47